Isolate remindBirthday and CLI inside dedicated packages.
