+++ README.md
@@ -148,4 +148,4 @@ This work is published under the MIT license.
 Please see the `LICENSE` file for details.
 
 * * *
-Automatically generated by [autoreadme](https://github.com/jimmyfrasche/autoreadme) on 2018.09.19
+Automatically generated by [autoreadme](https://github.com/jimmyfrasche/autoreadme) on 2018.09.20
